wip(reading): 划词共享层 + 保存为笔记流程 + PDF 接入（未完成）
