+++ src/shared/notes/index.ts
@@ -1,0 +1,15 @@
+export {
+  deriveNoteTitle,
+  openSavedNote,
+  saveTextAsNote,
+  saveTextAsNoteAndNotify,
+  notifySaveTextAsNoteResult,
+} from './saveTextAsNote';
+export type { SaveTextAsNoteInput, SaveTextAsNoteResult } from './saveTextAsNote';
+export { useSaveAsNoteFlow, SaveAsNoteFolderPicker } from './useSaveAsNoteFlow';
+export type {
+  SaveAsNoteFlow,
+  SaveAsNoteRequest,
+  SaveAsNoteFolderPickerProps,
+  UseSaveAsNoteFlowOptions,
+} from './useSaveAsNoteFlow';
